voice/pyproject.toml
[project]
name = "alive-voice"
version = "0.1.0"
description = "Norwegian voice pipeline — Smart Turn v3 + NB-Whisper + Claude"
requires-python = ">=3.10,<3.14"
dependencies = [
    "websockets>=14.0",
    "numpy>=1.24.0",
    "onnxruntime>=1.17.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "transformers>=4.45.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
